+++ alicompat/neoforge/src/compat/apotheosis/java/com/yanny/alicompat/compat/apotheosis/AffixLootPoolEntryAccessor.java
@@ -16,7 +16,7 @@ import com.yanny.alicompat.accessor.IEntryTooltip;
 import dev.shadowsoffire.apotheosis.loot.AffixLootEntry;
 import dev.shadowsoffire.apotheosis.loot.LootRarity;
 import dev.shadowsoffire.apotheosis.loot.entry.AffixLootPoolEntry;
-import dev.shadowsoffire.placebo.reload.DynamicHolder;
+import dev.shadowsoffire.placebo.dynreg.DynamicHolder;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.storage.loot.functions.LootItemFunction;
 import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;