+++ alicompat/neoforge/src/compat/apotheosis/java/com/yanny/alicompat/compat/apotheosis/ReforgeItemFunctionAccessor.java
@@ -8,7 +8,7 @@ import com.yanny.alicompat.accessor.FieldAccessor;
 import com.yanny.alicompat.accessor.IFunctionTooltip;
 import dev.shadowsoffire.apotheosis.loot.LootRarity;
 import dev.shadowsoffire.apotheosis.loot.functions.ReforgeItemFunction;
-import dev.shadowsoffire.placebo.reload.DynamicHolder;
+import dev.shadowsoffire.placebo.dynreg.DynamicHolder;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Set;