+++ alicompat/neoforge/src/compat/kaleidoscopecookery/java/com/yanny/alicompat/compat/kaleidoscopecookery/package-info.java
@@ -1,0 +1,4 @@
+@ParametersAreNonnullByDefault
+package com.yanny.alicompat.compat.kaleidoscopecookery;
+
+import javax.annotation.ParametersAreNonnullByDefault;